all instructions in rt.jar captured

+++ src/main/java/lk/ac/mrt/cse/mscresearch/util/IOUtil.java
@@ -5,11 +5,20 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.FileWriter;
 import java.io.IOException;
+import java.io.StringWriter;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
@@ -19,8 +28,10 @@ import org.apache.commons.exec.ExecuteException;
 import org.apache.commons.exec.PumpStreamHandler;
 import org.apache.commons.io.FileUtils;
 
-public class CommandLineUtil {
+public class IOUtil {
 
+	private static final Pattern hashIndexFileFormat = Pattern.compile("(?<hash>[0-9A-F]{32})=(?<class>.+)");
+	
 	public String disassembleClass(String className, String classpath){
 		CommandLine cmdLine = new CommandLine("javap");
 //		cmdLine.addArgument("-l");
@@ -75,12 +86,19 @@ public class CommandLineUtil {
 
 	public File copy(File src, File dest) throws IOException {
 		File copy  = new File(dest, src.getName());
-		FileUtils.copyFile(src, copy);
+		if(!copy.exists()){
+			FileUtils.copyFile(src, copy);
+		}
 		return copy;
 	}
 	
 	public Map<String, String> unzip(File jar) throws IOException {
 		File directory = jar.getParentFile();
+		File unzipMarker = new File(directory, "step.1.unzip.done");
+		File fileHashesSaved = new File(directory, "step.1.unzip.hashes");
+		if(unzipMarker.exists()) {
+			return loadFileHashes(fileHashesSaved);
+		}
         ZipInputStream zis = new ZipInputStream(new FileInputStream(jar));
         ZipEntry zipEntry = zis.getNextEntry();
         Map<String, String> fileHashes = new HashMap<>();
@@ -94,13 +112,46 @@ public class CommandLineUtil {
         }
         zis.closeEntry();
         zis.close();
+        saveHashes(fileHashesSaved, fileHashes);
+        unzipMarker.createNewFile();
 		return fileHashes;
+	}
+
+	private void saveHashes(File fileHashesSaved, Map<String, String> fileHashes) throws IOException {
+		try(FileWriter w = new FileWriter(fileHashesSaved, false)){
+			fileHashes.forEach((k,v)-> {
+				try {
+					w.write(v + "=" + v);
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			} );
+		}
+		
+	}
+
+	private Map<String, String> loadFileHashes(File fileHashesSaved) throws IOException {
+		List<String> s = getStrings(fileHashesSaved.getAbsolutePath());
+		Map<String, String> map = new HashMap<>();
+		s.stream().forEach(e->{
+			Matcher m = hashIndexFileFormat.matcher(e);
+			m.find();
+			map.put(m.group("class"), m.group("hash"));
+		});
+		return map;
 	}
 
 	protected void createFile(File diractory, ZipInputStream zis, ZipEntry zipEntry, Map<String, String> fileHashes) throws IOException, FileNotFoundException {
 		byte[] buffer = new byte[1024];
 		String name = zipEntry.getName();
 		File file = new File(diractory, name);
+		if(file.exists()) {
+			return;
+		}
+		File parent = file.getParentFile();
+		if(!parent.exists()) {
+			parent.mkdirs();
+		}
 		try(FileOutputStream fos = new FileOutputStream(file)){
 			int len;
 			while ((len = zis.read(buffer)) > 0) {
@@ -126,4 +177,49 @@ public class CommandLineUtil {
 		}
 		
 	}
+
+	public List<String> getStrings(File directory, String fileName) throws IOException {
+		return getStrings(directory.getAbsolutePath() + File.separatorChar + fileName);
+	}
+	
+	public List<String> getStrings(String filePath) throws IOException {
+		if(new File(filePath).exists()) {
+			return Files.readAllLines(Paths.get(filePath));
+		}
+		return Collections.emptyList();
+	}
+	
+	public String getAsString(String fileName) throws IOException {
+		List<String> decompiled = Files.readAllLines(Paths.get(fileName));
+		StringBuilder sb = new StringBuilder();
+		decompiled.stream().forEach(s->sb.append(s).append(","));
+		return sb.toString();
+	}
+	
+	public static void main(String[] ar) throws IOException {
+		IOUtil ioUtil = new IOUtil();
+		String s = ioUtil.getAsString("tmp");
+//		System.out.println(s);
+		String regEx = "[^\\d,:^ ^#]+";
+		Matcher m = Pattern.compile("[a-zA-Z]+\\.[a-zA-Z.$_\\d]+").matcher(s);
+		Set<String> flattened = new HashSet<>();
+		while(m.find()) {
+			flattened.add(s.substring(m.start(), m.end()));
+		}
+		File parent = new File("D:\\development\\msc-research\\Temp\\7A9E4B32BD67E7E16240D8B169298B8E");
+		List<String> current = ioUtil.getStrings(parent, "processed.list");
+		current.removeAll(flattened);
+		current.stream().forEach(System.out::println);
+//		System.out.println(flattened);
+		
+		try(FileWriter w = new FileWriter(new File(parent.getAbsolutePath() + File.separatorChar + "processed.list"), false)){
+			current.stream().forEach(
+					ss->{
+				try {
+					w.write(ss+"\r\n");
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			});
+	}}
 }
